quantum_inspired_lib/unified_quantum_nn.py
# quantum_inspired_lib/unified_quantum_nn.py

import numpy as np
from scipy.stats import unitary_group
from typing import List, Tuple
import logging
from logger import setup_logger

class UnifiedQuantumNeuron:
    def __init__(self, num_inputs: int):
        self.weights = unitary_group.rvs(num_inputs)
        self.bias = np.random.uniform(0, 2*np.pi)
        self.phase = np.random.uniform(0, 2*np.pi)

    def quantum_activation(self, x: np.ndarray) -> np.ndarray:
        return np.abs(np.cos(x + self.phase))**2 + 1j * np.abs(np.sin(x + self.phase))**2

    def forward(self, inputs: np.ndarray) -> np.ndarray:
        # Flatten all dimensions except the first (batch dimension)
        flattened_inputs = inputs.reshape(inputs.shape[0], -1)
        
        # Adjust weights if necessary
        if flattened_inputs.shape[1] != self.weights.shape[0]:
            self.weights = unitary_group.rvs(flattened_inputs.shape[1])
        
        z = np.dot(flattened_inputs, self.weights) + self.bias
        return self.quantum_activation(z)

class UnifiedQuantumLayer:
    def __init__(self, num_inputs: int, num_neurons: int, entanglement_strength: float = 0.1):
        self.neurons = [UnifiedQuantumNeuron(num_inputs) for _ in range(num_neurons)]
        self.num_neurons = num_neurons
        self.entanglement_size = num_neurons * 2
        self.entanglement_matrix = self.generate_entanglement_matrix(self.entanglement_size, entanglement_strength)

    def generate_entanglement_matrix(self, size: int, strength: float) -> np.ndarray:
        matrix = np.eye(size) + strength * np.random.randn(size, size)
        return matrix / np.linalg.norm(matrix, axis=1, keepdims=True)

    def forward(self, inputs: np.ndarray) -> np.ndarray:
        individual_outputs = np.array([neuron.forward(inputs) for neuron in self.neurons])
        
        # Reshape to (batch_size, num_neurons, -1)
        batch_size = inputs.shape[0]
        reshaped_outputs = individual_outputs.transpose(1, 0, 2)
        
        # Flatten the last dimension
        flattened_outputs = reshaped_outputs.reshape(batch_size, self.num_neurons, -1)
        
        # Adjust entanglement matrix if necessary
        if flattened_outputs.shape[2] * self.num_neurons != self.entanglement_size:
            self.entanglement_size = flattened_outputs.shape[2] * self.num_neurons
            self.entanglement_matrix = self.generate_entanglement_matrix(self.entanglement_size, 0.1)
        
        # Reshape for entanglement
        entanglement_input = flattened_outputs.reshape(batch_size, -1)
        
        # Apply entanglement
        entangled_outputs = np.matmul(entanglement_input, self.entanglement_matrix.T)
        
        # Reshape back to (batch_size, num_neurons, -1)
        return entangled_outputs.reshape(batch_size, self.num_neurons, -1)

class UnifiedQuantumNeuralNetwork:
    def __init__(self, layer_sizes: List[int], entanglement_strength: float = 0.1):
        self.logger = setup_logger(self.__class__.__name__, logging.INFO)
        self.layers = []
        for i in range(1, len(layer_sizes)):
            self.layers.append(UnifiedQuantumLayer(layer_sizes[i-1], layer_sizes[i], entanglement_strength))
        self.final_layer_size = layer_sizes[-1]

    def forward(self, inputs: np.ndarray) -> np.ndarray:
        self.logger.debug(f"Forward pass, input shape: {inputs.shape}")
        for i, layer in enumerate(self.layers):
            inputs = layer.forward(inputs)
            self.logger.debug(f"After layer {i+1}, shape: {inputs.shape}")
        # Ensure the output shape is (batch_size, final_layer_size)
        return inputs.reshape(-1, self.final_layer_size)

    def quantum_attention(self, inputs: np.ndarray) -> Tuple[np.ndarray, np.ndarray]:
        forward_output = self.forward(inputs)
        attention_weights = np.abs(forward_output)**2
        attention_weights /= np.sum(attention_weights, axis=1, keepdims=True)
        attended_output = np.sum(inputs.reshape(inputs.shape[0], -1)[:, :, np.newaxis] * attention_weights[:, np.newaxis, :], axis=2)
        return attended_output, attention_weights

    def quantum_backpropagation(self, X: np.ndarray, y: np.ndarray, learning_rate: float = 0.01, epochs: int = 100):
        for epoch in range(epochs):
            total_loss = 0
            for x, target in zip(X, y):
                x = x.reshape(1, -1)  # Ensure x is 2D
                target = target.reshape(1, -1)  # Ensure target is 2D
                output = self.forward(x)
                loss = np.mean(np.abs(output - target)**2)
                total_loss += loss

                # Simplified backpropagation (gradient descent)
                error = output - target
                for layer in reversed(self.layers):
                    # Update weights and biases (simplified)
                    layer.weights -= learning_rate * np.dot(layer.last_input.T, error)
                    layer.bias -= learning_rate * np.sum(error, axis=0)
                    # Compute error for the next layer
                    error = np.dot(error, layer.weights.T)

            if epoch % 10 == 0:
                self.logger.info(f"Epoch {epoch}, Loss: {total_loss / len(X)}")
